requirements.txt
firebase-admin
streamlit
altair
pandas
numpy
google-cloud-firestore
google-auth
